+++ script/core/diagnostics/missing-local-doc.lua
@@ -1,0 +1,27 @@
+local files  = require 'files'
+local guide  = require 'parser.guide'
+local await  = require 'await'
+local helper = require 'core.diagnostics.helper.missing-doc-helper'
+
+---@async
+return function (uri, callback)
+    local state = files.getState(uri)
+
+    if not state then
+        return
+    end
+
+    if not state.ast then
+        return
+    end
+
+    ---@async
+    guide.eachSourceType(state.ast, 'function', function (source)
+        await.delay()
+
+        if source.parent.type ~= 'local' then return end
+
+        helper.CheckFunction(source, callback, 'DIAG_MISSING_LOCAL_DOC_COMMENT', 'DIAG_MISSING_LOCAL_DOC_PARAM',
+            'DIAG_MISSING_LOCAL_DOC_RETURN')
+    end)
+end